refactor(slice-0): extract jwtbeareroptions setup to drop program.cs complexity
sonarcloud s3776 flagged program.cs cognitive complexity at 16/15 after
db6de3d's bridge extraction dropped it from 18. the jwtbeareroptions
`.Configure<IOptions<JwtAuthOptions>>((bearer, jwt) => { … })` lambda
carried the remaining over-quota branch (signingkey null-check +
keyid-null-coalesce + object initializer).

- new `infrastructure/jwtbeareroptionssetup.cs` implements
  `iconfigurenamedoptions<jwtbeareroptions>` and does the same work
  in two private statics (`buildtokenvalidationparameters` +
  `buildsigningkey`). strict-always posture + hs256-pin + mapinboundclaims
  = false are identical to the previous inline lambda.
- program.cs replaces the `.configure<…>` chain with a one-line
  `addsingleton<iconfigureoptions<jwtbeareroptions>, jwtbeareroptionssetup>()`
  registration. options framework invokes the setup on first resolution
  of `ioptionsmonitor<jwtbeareroptions>.Get("Bearer")`, same as before.

behavior identical — all 639 tests pass unchanged, including the new
me-with-non-guid-sub case 10c from db6de3d. sonarcloud complexity
should now read ≤ 15.
